refactor(logic files): change file names for clarity
