+++ scheduler/src/interfaces/server/agent_interface.rs
@@ -1,6 +1,6 @@
-use crate::logic::agent_logic::Agent as PoolAgent;
-use crate::logic::agent_logic::{compute_score, AgentPool};
-use crate::logic::agent_logic::Hostname;
+use crate::logic::agent_pool_logic::Agent as PoolAgent;
+use crate::logic::agent_pool_logic::{compute_score, AgentPool};
+use crate::logic::agent_pool_logic::Hostname;
 use log::{error, info};
 
 //use crate::proto::agent as proto;
