+++ scheduler/src/logic/mod.rs
@@ -1,2 +1,2 @@
-pub mod agent_logic;
-pub mod controller_logic;
+pub mod agent_pool_logic;
+pub mod action_queue_logic;
