+++ scheduler/src/main.rs
@@ -3,8 +3,8 @@ use tokio::sync::Mutex;
 
 use env_logger;
 use log::info;
-use logic::agent_logic::AgentPool;
-use logic::controller_logic::ActionsQueue;
+use logic::agent_pool_logic::AgentPool;
+use logic::action_queue_logic::ActionsQueue;
 use tonic::transport::Server;
 
 mod proto;
